+++ lib/generators/recommendable/install_generator.rb
@@ -3,10 +3,7 @@ require 'rails/generators'
 module Recommendable
   module Generators
     class InstallGenerator < Rails::Generators::Base
-      argument     :redis_host,   :type => :string,  :default => "localhost",       :desc => "The hostname your redis server is running on."
-      argument     :redis_port,   :type => :string,  :default => "6379",            :desc => "The port your redis server is running on."
-      class_option :redis_socket, :type => :string,                                 :desc => "Indicates the UNIX socket your redis server is running on (if it is)."
-      class_option :no_migrate,   :type => :boolean, :default => false,             :desc => "Skip migrations. The Like and Dislike tables will not be created."
+      class_option :no_migrate,   :type => :boolean, :default => false, :desc => "Skip migrations. The Like and Dislike tables will not be created."
 
       source_root File.expand_path("../templates", __FILE__)