+++ lib/generators/recommendable/templates/initializer.rb
@@ -1,28 +1,23 @@
 require "redis"
 
-# Recommendable requires a connection to a running redis-server. Either create
-# a new instance based on a host/port or UNIX socket, or pass in an existing
-# Redis client instance.
-<% if options.redis_socket %># <% end %>Recommendable.redis = Redis.new(:host => "<%= redis_host %>", :port => <%= redis_port %>)
+Recommendable.configure do |config|
+  # Recommendable requires a connection to a running redis-server. Create a new
+  # instance based on a host/port or UNIX socket, or pass in an existing client.
+  config.redis = Redis.new(:host => 'localhost', :port => 6379)
 
-# Connect to Redis via a UNIX socket instead
-<% unless options.redis_socket %># <% end %>Recommendable.redis = Redis.new(:sock => "<%= options.redis_socket %>")
+  # Tell Recommendable how long to wait until unnecessary keys are expired.
+  # These keys point to sets that Recommendable uses internally; expiration is
+  # to avoid a set being destroyed while another background worker is using it.
+  # You should set this value to how long it takes for a worker process to complete.
+  # Set to false if you do not wish to expire these keys. Set to :destroy if they
+  # can immediately be torn down (you do not have concurrent workers)
+  config.expire_keys_in = 1.hour
 
-# Tell Redis which database to use (usually between 0 and 15). The default of 0
-# is most likely okay unless you have another application using that database.
-# Recommendable.redis.select "0"
+  # Configure the name of your Sidekiq or Resque queue
+  config.queue_name = :recommendable
 
-# If using Sidekiq, uncomment the following or place it in config/initializers/sidekiq.rb:
-# Sidekiq.configure_server do |config|
-#   require 'sidekiq/middleware/server/unique_jobs'
-#   config.server_middleware do |chain|
-#     chain.add Sidekiq::Middleware::Server::UniqueJobs
-#   end
-# end
-#
-# Sidekiq.configure_client do |config|
-#   require 'sidekiq/middleware/client/unique_jobs'
-#   config.client_middleware do |chain|
-#     chain.add Sidekiq::Middleware::Client::UniqueJobs
-#   end
-# end
+  # Automatically enqueue users to have their recommendations refreshed after
+  # liking/disliking an object. To manually do this, you can use
+  # Recommendable.enqueue(user_id)
+  config.auto_enqueue = true
+end